+++ docs/changelog.md
@@ -35,6 +35,7 @@ As features stabilize some brief notes about them will accumulate here.
 * `gui-startup` event now also works with `wezterm ssh`
 * `x` and `+` buttons in the fancy tab bar are now always square [#2399](https://github.com/wez/wezterm/issues/2399)
 * middle clicking a tab to close it will now confirm closing using the same rules as [CloseCurrentTab](config/lua/keyassignment/CloseCurrentTab.md) [#2350](https://github.com/wez/wezterm/issues/2350)
+* Emitting the tmux-style `ESC k TITLE ST` sequence via ConPTY breaks output for the pane [#2442](https://github.com/wez/wezterm/issues/2442)
 
 ### 20220807-113146-c2fee766